Update DeepSeek tests to use Fixtures namespace and move helper to Helpers.php
